+++ lib/constants/app_constants.dart
@@ -2,4 +2,6 @@ class KContents {
   KContents._();
 
   static const String kAppName = "kroBanking";
+  static const String kProfileIcon =
+      "https://avatars.githubusercontent.com/u/35768100?v=4";
 }
